+++ lib/util/macos_utils.dart
@@ -28,4 +28,9 @@ class MacosUtils {
       'thread': thread,
     });
   }
+
+  /// Redirects `stdout` and `stderr` streams to a `app.log` file.
+  static Future<bool> redirectStdOut() async {
+    return await _platform.invokeMethod('redirectStdOut');
+  }
 }